requirements.txt
torch>=2.0.0
diffusers>=0.25.0
transformers>=4.30.0
open-clip-torch>=2.24.0
scikit-learn>=1.3.0
scipy>=1.10.0
numpy>=1.24.0
Pillow>=9.0.0
matplotlib>=3.7.0
tqdm>=4.60.0
pyyaml>=6.0
